fix(payload): address PR review on bounded SSZ decode
- Set MAX_WITHDRAWALS_PER_PAYLOAD to 0; Base requires the withdrawals
  list to be present but empty, so any non-empty list is rejected.
- Move the allocation-amplification regression test (and its crate-wide
  #[global_allocator]) into a dedicated integration test binary so the
  allocator is isolated from the rest of the crate's unit tests.
